refactor: move all common chart files into `charts` module
