+++ src/explore-education-statistics-admin/src/pages/release/edit-release/manage-datablocks/components/ChartTypeSelector.tsx
@@ -1,5 +1,5 @@
 import styles from '@admin/pages/release/edit-release/manage-datablocks/components/graph-builder.module.scss';
-import { ChartDefinition } from '@common/modules/find-statistics/components/charts/types/chart';
+import { ChartDefinition } from '@common/modules/charts/types/chart';
 import classnames from 'classnames';
 import React from 'react';
