+++ src/explore-education-statistics-common/src/modules/charts/components/MapBlock.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../src/styles/govuk-base';
+@import 'src/styles/govuk-base';
 
 .map {
   background-color: #f8f8fc;
