+++ src/explore-education-statistics-common/src/modules/charts/components/__tests__/MapBlock.test.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable */
-import testData from '@common/modules/find-statistics/components/charts/__tests__/__data__/testBlockData';
+import testData from '@common/modules/charts/components/__tests__/__data__/testBlockData';
 import React from 'react';
 
 import { render, wait } from 'react-testing-library';
-import MapBlock from '../MapBlock';
+import MapBlock from 'src/modules/charts/components/MapBlock';
 
 describe('MapBlock', () => {
   test('renders', async () => {
