+++ src/explore-education-statistics-common/src/modules/charts/components/__tests__/__data__/testBlockData.ts
@@ -2,8 +2,8 @@
 import {
   ChartMetaData,
   ChartProps,
-} from '@common/modules/find-statistics/components/charts/types/chart';
-import { parseMetaData } from '@common/modules/find-statistics/components/charts/util/chartUtils';
+} from '@common/modules/charts/types/chart';
+import { parseMetaData } from '@common/modules/charts/util/chartUtils';
 import {
   DataBlockData,
   DataBlockMetadata,
@@ -11,9 +11,9 @@ import {
   GeographicLevel,
 } from '@common/services/dataBlockService';
 
-import Features from './testLocationData';
+import Features from 'src/modules/charts/components/__tests__/__data__/testLocationData';
 
-import testResponseData_23_26_28__1_2_LA_JSON from './testResponseData_23_26_28__1_2_LA.json';
+import testResponseData_23_26_28__1_2_LA_JSON from 'src/modules/charts/components/__tests__/__data__/testResponseData_23_26_28__1_2_LA.json';
 
 const testResponseData_23_26_28__1_2_LA: DataBlockResponse = (testResponseData_23_26_28__1_2_LA_JSON as unknown) as DataBlockResponse;
