+++ src/explore-education-statistics-common/src/modules/charts/components/__tests__/__data__/testLocationData.ts
