+++ src/explore-education-statistics-common/src/modules/charts/components/charts.scss
