+++ src/explore-education-statistics-common/src/modules/find-statistics/components/__tests__/DataBlock.test.tsx
@@ -1,4 +1,4 @@
-import testData from '@common/modules/find-statistics/components/charts/__tests__/__data__/testBlockData';
+import testData from '@common/modules/charts/components/__tests__/__data__/testBlockData';
 import _dataBlockService, {
   DataBlockRequest,
   GeographicLevel,
