+++ src/explore-education-statistics-common/src/services/__tests__/dataBlockService.test.ts
@@ -3,7 +3,7 @@ import dataBlockService, {
   DataBlockRequest,
   GeographicLevel,
 } from '@common/services/dataBlockService';
-import testData from '@common/modules/find-statistics/components/charts/__tests__/__data__/testBlockData';
+import testData from '@common/modules/charts/components/__tests__/__data__/testBlockData';
 
 import { dataApi } from '@common/services/api';
